Untokenize PillButtonBar
